fix the module path
The module was defined as github.com/ipld/go-codec-dagpb/dagpb, but the
repository is at github.com/ipld/go-codec-dagpb. This means that neither
can be resolved as a valid Go package import path; the former results in
a GitHub 404, and the latter results in a valid GitHub repository which
holds a different module.

Rod confirmed that the /dagpb suffix wasn't intentional, so just get rid
of it.

While at it, mod tidy.