add a prototype chooser API for integrations
At least go-merkledag and go-graphsync still need a prototype chooser,
and go-ipld-prime-proto provides one, so this simplifies integration
with downstreams.

It's worth noting that we just add PBNode here, while
go-ipld-prime-proto also added RawNode. This is because this dag-pb
module leaves raw nodes out, since those are now in
go-ipld-prime/codec/raw.

The logic is just three lines, but it's still easier to not have to
copy-paste these between different projects.