bump to v1.4.0 given that we updated ipld-prime
We released a v1.3.1 which upgraded a major dependency, go-ipld-prime,
which is known for breaking some downstreams from time to time.
That should not have been a bugfix release, so we tagged v1.3.2 without
that version update.

To make sure that someone updating to the latest version still gets the
update, we're now doing v1.4.0 with the ipld-prime version update.